Split off funding-related states into separate file
In this commit, I've simply moved states that are related to the source
of funding, and how much of it there is, into its own file.

Tests still pass, but the intention is that
* states in the `redux/states/channels/funding/funding.ts` will be of type
`FundingState`
* states in `redux/states/channels/funding/index.ts` will be of type `FundingChannelState`
* A `ChannelState` may be a `FundingChannelState`
* a `FundingChannelState` has a `fundingState` attribute of type
`FundingState`

With the proposed change above, "funding" becomes a bit overloaded,
refering to the stage of the channel state as well as a property of
a channel state that is in the funding stage. We could perhaps rename
the "funding" stage accordingly.
